+++ src/lib.rs
@@ -46,6 +46,7 @@ pub mod osc;
 pub mod prelude;
 pub mod rand;
 pub mod state;
+pub mod time;
 pub mod ui;
 pub mod window;