+++ src/state.rs
@@ -282,91 +282,12 @@ pub mod mouse {
 
 /// Tracked durations related to the App.
 pub mod time {
-    use std::{ops, time};
-    use std::cmp::Ordering;
-
     /// The state of time tracked by the App.
     #[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Hash)]
     pub struct Time {
         /// The duration since the app started running.
-        pub since_start: Duration,
+        pub since_start: std::time::Duration,
         /// The duration since the previous update.
-        pub since_prev_update: Duration,
-    }
-
-    /// A wrapper around a std duration with simpler methods.
-    #[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Hash, Ord, PartialOrd)]
-    pub struct Duration {
-        /// The inner std duration.
-        pub duration: time::Duration,
-    }
-
-    impl From<time::Duration> for Duration {
-        fn from(duration: time::Duration) -> Self {
-            Duration { duration }
-        }
-    }
-
-    impl Into<time::Duration> for Duration {
-        fn into(self) -> time::Duration {
-            self.duration
-        }
-    }
-
-    impl ops::Deref for Duration {
-        type Target = time::Duration;
-        fn deref(&self) -> &Self::Target {
-            &self.duration
-        }
-    }
-
-    impl ops::DerefMut for Duration {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.duration
-        }
-    }
-
-    impl PartialEq<time::Duration> for Duration {
-        fn eq(&self, other: &time::Duration) -> bool {
-            self.duration == *other
-        }
-    }
-
-    impl PartialOrd<time::Duration> for Duration {
-        fn partial_cmp(&self, other: &time::Duration) -> Option<Ordering> {
-            self.duration.partial_cmp(other)
-        }
-    }
-
-    impl Duration {
-        /// A simple way of retrieving the duration as weeks.
-        pub fn weeks(&self) -> f64 {
-            self.days() / 7.0
-        }
-
-        /// A simple way of retrieving the duration as days.
-        pub fn days(&self) -> f64 {
-            self.hrs() / 24.0
-        }
-
-        /// A simple way of retrieving the duration as hrs.
-        pub fn hrs(&self) -> f64 {
-            self.mins() / 60.0
-        }
-
-        /// A simple way of retrieving the duration as minutes.
-        pub fn mins(&self) -> f64 {
-            self.secs() / 60.0
-        }
-
-        /// A simple way of retrieving the duration in seconds.
-        pub fn secs(&self) -> f64 {
-            self.as_secs() as f64 + self.subsec_nanos() as f64 * 1e-9
-        }
-
-        /// A simple way of retrieving the duration in milliseconds.
-        pub fn ms(&self) -> f64 {
-            self.secs() * 1_000.0
-        }
+        pub since_prev_update: std::time::Duration,
     }
 }